samples/sem-images-stats/requirements.txt
jupyter
numpy
matplotlib
scipy
pytesseract
pillow
toolz
pandas
scikit-image
dask